feat: rearrange entries
Add an order state variable to maintain an array
which specifies the order of the entries using
the id of the entries.

Add two functions, one each to reorder an entry
up by one position, and down by one position.

Add updater function to maintain this behaviour
for the existing entry creation/deletion
features.