feat: add skills and projects sections
The 'Skills' and 'Projects' sections will use
their own customiser and viewer components, as
their input fields differ from that of the
'Education' and 'Experience' sections quite a bit.

Splitting them up like this helps to avoid having
an excessive number of optional fields in the
HTML forms, and conditional rendering is kept
under control for better clarity.