refactor: unify form input states into an object
Use a `form` object state with a `handleChange`
handler function which calls on the `setForm`
updater function to keep all input values in the
form up to date.

Exception is made for the mobileInput as the
PhoneInput component from react-phone-number-input
is set up in a way where its onChange must point
to its own updater function for the country
selector function to work.