fix: add flex-grow styling to viewer components
Since the 3 Viewer subcomponents are flexboxes,
and flex-direction defaults to 'row', the content
of each flex item only takes up as much horizontal
space as it needs.

As such, the ViewerGeneral section shown in
issue #3 does not take up full width, since its
content does not actually fill the full width.

This dependence on the content length is
eliminated by setting `flex-grow: 1` on the
section container.