fix: use empty string instead of null
Swap out all state variables' initial default
value from null to empty string '' for `useState`
hooks that do not deal with Date objects.

Do the same for when the associated updater
functions are called for form resets.

Exception made for state variables dealing with
Date objects as those are properly dealt with
by the dayjs library.